+++ tests/testthat/test-properties.R
@@ -363,3 +363,28 @@ test_that("AllVariablesNames functions in formula",
     out <- AllVariablesNames(log(y) ~ I(log(x))+ `a(b`)
     expect_equal(out, c("y", "x", "`a(b`"))
 })
+
+test_that("OutcomeName function in response",
+{
+    out <- OutcomeName(log(response) ~ I(log(x))+ `a(b`)
+    expect_equal(out, "response")
+})
+
+test_that("OutcomeName I() in response",
+{
+    out <- OutcomeName(I(y^2) ~ I(log(x))*z)
+    expect_equal(out, "y")
+})
+
+test_that("OutcomeName backticks in response",
+{
+    out <- OutcomeName(`dat$var$y` ~ I(log(x))+ `a(b`)
+    expect_equal(out, "`dat$var$y`")
+})
+
+test_that("OutcomeName response has $ without backticks",
+{
+    out <- OutcomeName(dat$y ~ .)
+    expect_equal(out, "dat$y")
+})
+