+++ inst/NEWS.md
@@ -1,12 +1,12 @@
-Changes in Version 1.0.5 (2017-11-01)
+Changes in Version 1.0.6 (2017-11-02)
 --------------------------------------------------------
 
 BUG FIXES
 
 * `AllVariablesNames` now correctly handles interactions
 in formulae (DS-1527)
-* `AllVariablesNames` now correctly handles function calls
-in formulae, e.g. `y~log(x)+I(x^2)` (DS-1527)
+* `AllVariablesNames` and `OutcomeName`now correctly 
+handle function calls in formulae, e.g. `y~log(x)+I(x^2)` (DS-1527)
 
 Changes in Version 1.0.3 (2017-10-31)
 --------------------------------------------------------